// src/types/enums/ErrorMessage.ts
export enum ErrorMessage {
  EMAIL_REQUIRED = 'Email is required',
  PASSWORD_REQUIRED = 'Password is required',
  ALL_FIELDS_REQUIRED = 'All fields are required',
  EMAIL_ALREADY_EXISTS = 'Email already exists',
  INVALID_CREDENTIALS = 'Invalid email or password',
  USER_NOT_FOUND = 'User not found',
  INTERNAL_SERVER_ERROR = 'Internal server error',
  OTP_REQUIRED = 'OTP is required',
  INVALID_OTP = 'Invalid or expired OTP',
  TOKEN_REQUIRED = 'Token is required',
  INVALID_TOKEN = 'Invalid token',
  REFRESH_TOKEN_REQUIRED = 'Refresh token is required',
  INVALID_REFRESH_TOKEN = 'Invalid refresh token',
  REFRESH_TOKEN_EXPIRED = 'Refresh token expired',
  ACCOUNT_BLOCKED = 'Your account has been blocked. Please contact admin for support.',
  CURRENT_PASSWORD_INCORRECT = 'Current password is incorrect',
  PASSWORD_UPDATE_FAILED = 'Failed to update password',
  GOOGLE_TOKEN_REQUIRED = 'Google token is required',
  INVALID_GOOGLE_TOKEN = 'Invalid Google token',
  USER_ID_REQUIRED = 'User ID is required',
  DRIVER_NOT_FOUND = 'Driver not found',
  LOGOUT_SUCCESS = 'Logged out successfully',
  PASSWORD_RESET_SUCCESS = 'Password reset successful',
  EMAIL_UPDATE_SUCCESS = 'Email updated successfully',
  OTP_SEND_SUCCESS = 'OTP sent successfully',
  OTP_VERIFY_SUCCESS = 'OTP verified successfully',
  LOGIN_SUCCESS = 'Login successful',
  REGISTRATION_SUCCESS = 'Registration successful',
  DRIVER_REGISTRATION_SUCCESS = 'Driver email registered successfully',
  PASSWORD_UPDATE_SUCCESS = 'Password updated successfully'
} 